pyproject.toml
[tool.poetry]
name = "tweet_app"
version = "0.1.0"
description = ""
authors = ["LionelB"]

[tool.poetry.dependencies]
python = "^3.8"
pandas = "^1.4.1"
tweepy = "3.9.0"
emoji = "^1.7.0"
black = "^22.3.0"
isort = "^5.10.1"
supervisor = "^4.2.4"
s3fs = "^2022.3.0"
fsspec = "^2022.3.0"
python-decouple = "^3.6"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
